refactored list of tabs to be SideNavTabsList component
